requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
sentence-transformers
groq
requests
pandas
fastapi[all]
openai